refactor: ctr package
