+++ internal/ctr/cgroups.go
@@ -20,7 +20,6 @@ import (
 	"bufio"
 	"context"
 	"errors"
-	"log/slog"
 	"os"
 	"path/filepath"
 	"strings"
@@ -31,73 +30,6 @@ import (
 )
 
 // TODO(eminwux): add cgroup integration tests once CI exposes a writable cgroup v2 hierarchy.
-
-var (
-	errEmptyGroupPath   = errors.New("ctr: cgroup group path is required")
-	errInvalidPID       = errors.New("ctr: pid must be greater than zero")
-	errInvalidCPUWeight = errors.New("ctr: cpu weight must be within [1, 10000]")
-	errInvalidIOWeight  = errors.New("ctr: io weight must be within [1, 1000]")
-	errInvalidThrottle  = errors.New("ctr: io throttle entries require type, major, minor and rate")
-)
-
-// CgroupSpec describes how to create a new cgroup.
-type CgroupSpec struct {
-	// Group is the target cgroup path, e.g. /kukeon/workloads/runner.
-	Group string
-	// Mountpoint overrides the default cgroup mount (/sys/fs/cgroup) when non-empty.
-	Mountpoint string
-	// Resources defines the controller knobs that should be configured for the cgroup.
-	Resources CgroupResources
-}
-
-// CgroupResources represents the subset of controllers we expose.
-type CgroupResources struct {
-	CPU    *CPUResources
-	Memory *MemoryResources
-	IO     *IOResources
-}
-
-// CPUResources maps to cpu*, cpuset* controllers.
-type CPUResources struct {
-	Weight *uint64
-	Quota  *int64
-	Period *uint64
-	Cpus   string
-	Mems   string
-}
-
-// MemoryResources maps to memory controller knobs.
-type MemoryResources struct {
-	Min  *int64
-	Max  *int64
-	Low  *int64
-	High *int64
-	Swap *int64
-}
-
-// IOResources exposes IO weight + throttling.
-type IOResources struct {
-	Weight   uint16
-	Throttle []IOThrottleEntry
-}
-
-// IOThrottleType identifies the throttle file to target.
-type IOThrottleType string
-
-const (
-	IOTypeReadBPS   IOThrottleType = IOThrottleType(cgroup2.ReadBPS)
-	IOTypeWriteBPS  IOThrottleType = IOThrottleType(cgroup2.WriteBPS)
-	IOTypeReadIOPS  IOThrottleType = IOThrottleType(cgroup2.ReadIOPS)
-	IOTypeWriteIOPS IOThrottleType = IOThrottleType(cgroup2.WriteIOPS)
-)
-
-// IOThrottleEntry represents a single io.max entry.
-type IOThrottleEntry struct {
-	Type  IOThrottleType
-	Major int64
-	Minor int64
-	Rate  uint64
-}
 
 // NewCgroup provisions a new cgroup for the provided spec.
 func (c *client) NewCgroup(spec CgroupSpec) (*cgroup2.Manager, error) {
@@ -209,7 +141,7 @@ func (c *client) UpdateCgroup(group, mountpoint string, resources CgroupResource
 // AddProcessToCgroup adds the pid into cgroup.procs.
 func (c *client) AddProcessToCgroup(group, mountpoint string, pid int) error {
 	if pid <= 0 {
-		return errInvalidPID
+		return ErrInvalidPID
 	}
 	manager, err := c.managerFor(group, mountpoint)
 	if err != nil {
@@ -255,276 +187,6 @@ func (c *client) CgroupMetrics(group, mountpoint string) (*cgroupstats.Metrics, 
 		return nil, err
 	}
 	return manager.Stat()
-}
-
-func (r CgroupResources) toResources() (*cgroup2.Resources, error) {
-	resources := &cgroup2.Resources{}
-
-	if r.CPU != nil && !r.CPU.isZero() {
-		cpu, err := r.CPU.toResource()
-		if err != nil {
-			return nil, err
-		}
-		resources.CPU = cpu
-	}
-
-	if r.Memory != nil && !r.Memory.isZero() {
-		resources.Memory = r.Memory.toResource()
-	}
-
-	if r.IO != nil && !r.IO.isZero() {
-		io, err := r.IO.toResource()
-		if err != nil {
-			return nil, err
-		}
-		resources.IO = io
-	}
-
-	return resources, nil
-}
-
-func (c *CPUResources) isZero() bool {
-	if c == nil {
-		return true
-	}
-	return c.Weight == nil &&
-		c.Quota == nil &&
-		c.Period == nil &&
-		c.Cpus == "" &&
-		c.Mems == ""
-}
-
-func (c *CPUResources) toResource() (*cgroup2.CPU, error) {
-	if c.Weight != nil {
-		if *c.Weight < 1 || *c.Weight > 10000 {
-			return nil, errInvalidCPUWeight
-		}
-	}
-	cpu := &cgroup2.CPU{
-		Weight: c.Weight,
-		Cpus:   c.Cpus,
-		Mems:   c.Mems,
-	}
-	if c.Quota != nil || c.Period != nil {
-		cpu.Max = cgroup2.NewCPUMax(c.Quota, c.Period)
-	}
-	return cpu, nil
-}
-
-func (m *MemoryResources) isZero() bool {
-	if m == nil {
-		return true
-	}
-	return m.Min == nil &&
-		m.Max == nil &&
-		m.Low == nil &&
-		m.High == nil &&
-		m.Swap == nil
-}
-
-func (m *MemoryResources) toResource() *cgroup2.Memory {
-	return &cgroup2.Memory{
-		Min:  m.Min,
-		Max:  m.Max,
-		Low:  m.Low,
-		High: m.High,
-		Swap: m.Swap,
-	}
-}
-
-func (io *IOResources) isZero() bool {
-	if io == nil {
-		return true
-	}
-	return io.Weight == 0 && len(io.Throttle) == 0
-}
-
-func (io *IOResources) toResource() (*cgroup2.IO, error) {
-	if io.Weight != 0 && (io.Weight < 1 || io.Weight > 1000) {
-		return nil, errInvalidIOWeight
-	}
-	maxEntries := make([]cgroup2.Entry, 0, len(io.Throttle))
-	for _, entry := range io.Throttle {
-		if entry.Type == "" || entry.Rate == 0 || entry.Major < 0 || entry.Minor < 0 {
-			return nil, errInvalidThrottle
-		}
-		maxEntries = append(maxEntries, cgroup2.Entry{
-			Type:  cgroup2.IOType(entry.Type),
-			Major: entry.Major,
-			Minor: entry.Minor,
-			Rate:  entry.Rate,
-		})
-	}
-	return &cgroup2.IO{
-		BFQ: cgroup2.BFQ{Weight: io.Weight},
-		Max: maxEntries,
-	}, nil
-}
-
-func validateGroupPath(group string) error {
-	if group == "" {
-		return errEmptyGroupPath
-	}
-	return cgroup2.VerifyGroupPath(group)
-}
-
-func parseMountinfo(ctx context.Context, logger *slog.Logger) (string, error) {
-	// Read /proc/self/mountinfo to find cgroup2 mount point
-	if logger != nil {
-		logger.DebugContext(ctx, "parsing /proc/self/mountinfo to find cgroup2 mount")
-	}
-	file, err := os.Open("/proc/self/mountinfo")
-	if err != nil {
-		if logger != nil {
-			logger.DebugContext(ctx, "failed to open /proc/self/mountinfo", "error", err)
-		}
-		return "", err
-	}
-	defer file.Close()
-
-	scanner := bufio.NewScanner(file)
-	for scanner.Scan() {
-		line := scanner.Text()
-		// Format: space-separated fields
-		// Fields: mountID parentID major:minor root mountpoint mountopts - fstype source superopts
-		// Example: 36 35 98:0 /mnt1 /mnt2 rw,noatime master:1 - ext3 /dev/root rw,errors=continue
-		// We need to find the "-" separator, then fstype is the next field
-		fields := strings.Fields(line)
-		const minFieldsForMountinfo = 10
-		const mountpointFieldIndex = 4
-		if len(fields) < minFieldsForMountinfo {
-			continue
-		}
-
-		// Find the "-" separator (optional fields marker)
-		separatorIdx := -1
-		for i, field := range fields {
-			if field == "-" {
-				separatorIdx = i
-				break
-			}
-		}
-		if separatorIdx == -1 || separatorIdx+1 >= len(fields) {
-			continue
-		}
-
-		// Filesystem type is the field after "-"
-		fstype := fields[separatorIdx+1]
-		if fstype == "cgroup2" {
-			// Mount point is field 4 (0-indexed: fields[4])
-			if len(fields) > mountpointFieldIndex {
-				mountpoint := fields[mountpointFieldIndex]
-				if logger != nil {
-					logger.DebugContext(ctx, "found cgroup2 mount in mountinfo", "mountpoint", mountpoint)
-				}
-				return mountpoint, nil
-			}
-		}
-	}
-
-	if scanErr := scanner.Err(); scanErr != nil {
-		if logger != nil {
-			logger.DebugContext(ctx, "error scanning /proc/self/mountinfo", "error", scanErr)
-		}
-		return "", scanErr
-	}
-
-	if logger != nil {
-		logger.DebugContext(ctx, "cgroup2 mount not found in mountinfo")
-	}
-	return "", errors.New("cgroup2 mount not found in mountinfo")
-}
-
-func discoverCgroupMountpoint(ctx context.Context, logger *slog.Logger) (string, error) {
-	const fallbackMountpoint = consts.CgroupFilesystemPath
-
-	if logger != nil {
-		logger.DebugContext(ctx, "discovering cgroup mountpoint from /proc/self/cgroup")
-	}
-
-	// Read /proc/self/cgroup to find the cgroup2 entry
-	file, err := os.Open("/proc/self/cgroup")
-	if err != nil {
-		if logger != nil {
-			logger.DebugContext(
-				ctx,
-				"failed to open /proc/self/cgroup, using fallback",
-				"error",
-				err,
-				"fallback",
-				fallbackMountpoint,
-			)
-		}
-		// Fallback to default if we can't read the file
-		return fallbackMountpoint, nil
-	}
-	defer file.Close()
-
-	var cgroupPath string
-	scanner := bufio.NewScanner(file)
-	for scanner.Scan() {
-		line := scanner.Text()
-		// Format: hierarchy:controllers:path
-		// For cgroup2 (unified hierarchy): 0::<path> where controllers is empty
-		parts := strings.Split(line, ":")
-		if len(parts) == 3 && parts[0] == "0" && parts[1] == "" {
-			cgroupPath = parts[2]
-			if logger != nil {
-				logger.DebugContext(ctx, "found cgroup2 entry in /proc/self/cgroup", "cgroup_path", cgroupPath)
-			}
-			break
-		}
-	}
-
-	if scanErr := scanner.Err(); scanErr != nil {
-		if logger != nil {
-			logger.DebugContext(
-				ctx,
-				"error scanning /proc/self/cgroup, using fallback",
-				"error",
-				scanErr,
-				"fallback",
-				fallbackMountpoint,
-			)
-		}
-		// Fallback to default on scan error
-		return fallbackMountpoint, nil
-	}
-
-	// If we didn't find a cgroup2 entry, fallback to default
-	if cgroupPath == "" {
-		if logger != nil {
-			logger.DebugContext(
-				ctx,
-				"no cgroup2 entry found in /proc/self/cgroup, using fallback",
-				"fallback",
-				fallbackMountpoint,
-			)
-		}
-		return fallbackMountpoint, nil
-	}
-
-	// Get cgroup2 mount point from mountinfo
-	mountpoint, mountErr := parseMountinfo(ctx, logger)
-	if mountErr != nil {
-		if logger != nil {
-			logger.DebugContext(
-				ctx,
-				"failed to find cgroup2 mountpoint in mountinfo, using fallback",
-				"error",
-				mountErr,
-				"fallback",
-				fallbackMountpoint,
-			)
-		}
-		// Fallback to default if we can't find the mountpoint
-		return fallbackMountpoint, nil
-	}
-
-	if logger != nil {
-		logger.DebugContext(ctx, "discovered cgroup mountpoint", "mountpoint", mountpoint, "cgroup_path", cgroupPath)
-	}
-	return mountpoint, nil
 }
 
 func (c *client) effectiveMountpoint(mountpoint string) string {
@@ -612,4 +274,11 @@ func (c *client) managerFor(group, mountpoint string) (*cgroup2.Manager, error) 
 	}
 	c.storeManager(group, manager)
 	return manager, nil
+}
+
+func validateGroupPath(group string) error {
+	if group == "" {
+		return ErrEmptyGroupPath
+	}
+	return cgroup2.VerifyGroupPath(group)
 }
