+++ internal/ctr/namespaces.go
@@ -28,6 +28,54 @@ import (
 	"github.com/opencontainers/go-digest"
 )
 
+// namespaceCtx returns a context with the namespace set.
+func (c *client) namespaceCtx() context.Context {
+	c.namespaceMu.RLock()
+	defer c.namespaceMu.RUnlock()
+	return namespaces.WithNamespace(c.ctx, c.namespace)
+}
+
+// SetNamespace sets the namespace for subsequent operations.
+// This clears any previously set registry credentials.
+func (c *client) SetNamespace(namespace string) {
+	c.namespaceMu.Lock()
+	defer c.namespaceMu.Unlock()
+	c.namespace = namespace
+	c.logger.DebugContext(c.ctx, "set namespace", "namespace", namespace)
+
+	// Clear credentials when namespace is set without credentials
+	c.registryCredentialsMu.Lock()
+	defer c.registryCredentialsMu.Unlock()
+	c.registryCredentials = nil
+}
+
+// SetNamespaceWithCredentials sets the namespace and associated registry credentials.
+// This should be called when switching to a realm's namespace.
+func (c *client) SetNamespaceWithCredentials(namespace string, creds []RegistryCredentials) {
+	c.namespaceMu.Lock()
+	defer c.namespaceMu.Unlock()
+	c.namespace = namespace
+	c.logger.DebugContext(c.ctx, "set namespace with credentials", "namespace", namespace, "creds_count", len(creds))
+
+	c.registryCredentialsMu.Lock()
+	defer c.registryCredentialsMu.Unlock()
+	c.registryCredentials = creds
+}
+
+// GetRegistryCredentials returns the current registry credentials for the namespace.
+func (c *client) GetRegistryCredentials() []RegistryCredentials {
+	c.registryCredentialsMu.RLock()
+	defer c.registryCredentialsMu.RUnlock()
+	return c.registryCredentials
+}
+
+// Namespace returns the current namespace.
+func (c *client) Namespace() string {
+	c.namespaceMu.RLock()
+	defer c.namespaceMu.RUnlock()
+	return c.namespace
+}
+
 func (c *client) CreateNamespace(namespace string) error {
 	c.logger.DebugContext(c.ctx, "creating namespace", "namespace", namespace)
 	namespaces := c.cClient.NamespaceService()
@@ -141,7 +189,7 @@ func (c *client) GetNamespace(namespace string) (string, error) {
 // CleanupNamespaceResources removes all images and snapshots from a namespace
 // This must be called before deleting the namespace, as containerd requires
 // namespaces to be empty before deletion.
-func (c *client) CleanupNamespaceResources(ctx context.Context, namespace, snapshotter string) error {
+func (c *client) CleanupNamespaceResources(namespace, snapshotter string) error {
 	// Ensure client is connected
 	if c.cClient == nil {
 		if err := c.Connect(); err != nil {
@@ -150,7 +198,7 @@ func (c *client) CleanupNamespaceResources(ctx context.Context, namespace, snaps
 	}
 
 	// Set namespace context for operations
-	nsCtx := namespaces.WithNamespace(ctx, namespace)
+	nsCtx := namespaces.WithNamespace(c.ctx, namespace)
 
 	// Clean up images
 	c.logger.DebugContext(c.ctx, "listing images in namespace", "namespace", namespace)
